refactor: move termcode from internal to public API
Make termcode a public API available as aibo.termcode to provide stable,
supported access for users creating custom terminal key mappings.

Changes:
- Moved lua/aibo/internal/termcode.lua to lua/aibo/termcode.lua
- Exposed as aibo.termcode in the main module
- Updated all internal references to use aibo.termcode
- Updated documentation to show aibo.termcode.resolve() usage
- Added termcode.resolve to the public API documentation

This change acknowledges that termcode is essential functionality that
users need when creating custom mappings, and should therefore be part
of the stable public API rather than an internal implementation detail.

Users can now use:
  local aibo = require('aibo')
  aibo.send(aibo.termcode.resolve('<Up>'), bufnr)

Instead of reaching into internal modules.
