+++ doc/aibo.txt
@@ -670,40 +670,39 @@ Example usage:
 CUSTOM MAPPINGS					*aibo-custom-mappings*
 
 When creating custom mappings that send keys to the terminal, you MUST use
-`termcode.resolve()` instead of `vim.api.nvim_replace_termcodes()`.
+`aibo.termcode.resolve()` instead of `vim.api.nvim_replace_termcodes()`.
 
 						*aibo-termcode*
-Why use termcode.resolve()~
+Why use aibo.termcode.resolve()~
 
 The built-in `nvim_replace_termcodes()` returns Neovim's internal key
 representations, which terminal programs cannot understand. The
-`termcode.resolve()` function converts Vim key notation to actual ANSI
+`aibo.termcode.resolve()` function converts Vim key notation to actual ANSI
 escape sequences that terminals expect.
 
 Example difference:
 - nvim_replace_termcodes("<Up>") returns "\x80\x6B\x75" (internal code)
-- termcode.resolve("<Up>") returns "\27[A" (ANSI escape sequence)
+- aibo.termcode.resolve("<Up>") returns "\27[A" (ANSI escape sequence)
 
 						*aibo-custom-mapping-example*
 Creating custom key mappings~
 
 >
-	local termcode = require("aibo.internal.termcode")
 	local aibo = require("aibo")
 
 	-- Send navigation keys to terminal
 	vim.keymap.set('n', '<leader>au', function()
-	  aibo.send(termcode.resolve("<Up>"), bufnr)
+	  aibo.send(aibo.termcode.resolve("<Up>"), bufnr)
 	end, { buffer = bufnr, desc = "Send Up arrow" })
 
 	-- Send control sequences
 	vim.keymap.set('n', '<leader>ac', function()
-	  aibo.send(termcode.resolve("<C-c>"), bufnr)
+	  aibo.send(aibo.termcode.resolve("<C-c>"), bufnr)
 	end, { buffer = bufnr, desc = "Interrupt process" })
 
 	-- Send multiple keys
 	vim.keymap.set('n', '<leader>ah', function()
-	  local keys = termcode.resolve("<Home><S-End>")
+	  local keys = aibo.termcode.resolve("<Home><S-End>")
 	  aibo.send(keys, bufnr)
 	end, { buffer = bufnr, desc = "Select to end of line" })
 <
