+++ lua/aibo/init.lua
@@ -108,8 +108,7 @@ function M.submit(data, bufnr)
   aibo.controller.send(data)
 
   -- Convert submit key to terminal codes
-  local termcode = require("aibo.internal.termcode")
-  local submit_key = termcode.resolve("<CR>")
+  local submit_key = M.termcode.resolve("<CR>")
 
   vim.defer_fn(function()
     aibo.controller.send(submit_key)
@@ -119,5 +118,8 @@ function M.submit(data, bufnr)
     aibo.follow()
   end, config.submit_delay)
 end
+
+-- Expose termcode as part of the public API
+M.termcode = require("aibo.termcode")
 
 return M
